requirements.txt
lxml
podgen
requests
feedparser
beautifulsoup4
readability-lxml